back-end/python/requirements.txt
click==8.1.3
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
marshmallow==3.19.0
mysql-connector==2.2.9
mysqlclient==2.1.1
packaging==23.0
python-dotenv==0.21.1
SQLAlchemy==2.0.0
typing_extensions==4.4.0
Werkzeug==2.2.2
